requirements.txt
python-dotenv==0.17.1

aiomysql>=0.0.22
pytest>=7.3.1
loguru>=0.7.0
dateparser>=1.1.8